% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Chromosome-class}
\alias{Chromosome-class}
\alias{chromosome}
\alias{chromosome,ANY-method}
\title{Class \linkS4class{Chromosome}}
\description{
Class \linkS4class{Chromosome}

Constructor generic for \linkS4class{Chromosome}

Constructor method for \linkS4class{Chromosome}
}
\section{Slots}{

\describe{
\item{\code{Data}}{A \link{data.frame} with 6 columns (\code{"SNP"}, \code{"PVALUE"}, \code{"CHR"}, \code{"MAF"}, \code{"eSNP"}, \code{"xSNP"}).
Where \code{"eSNP"} and \code{"xSNP"} are logical columns defining the lists of SNPs (extended or not).}

\item{\code{LD}}{A data.frame which contains LD informations between SNPs (computed with \link{writeLD} or PLINK).}

\item{\code{eSNP}}{Contain a \linkS4class{EnrichSNP} object for a list of SNPs (eSNP).}

\item{\code{xSNP}}{Contain a \linkS4class{EnrichSNP} object for a extended list of SNPs (xSNP).}
}}

