% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareEnrichment.R
\name{compareEnrichment}
\alias{compareEnrichment}
\title{Compare enrichment analysis between two SNPs list}
\usage{
compareEnrichment(object.x, object.y, pattern = "Chrom", nSample = 100,
  empiricPvalue = TRUE, mc.cores = 1, onlyGenome = TRUE)
}
\arguments{
\item{object.x, object.y}{An \link{Enrichment} object fully filled (\emph{e.g.}, \link{readEnrichment}).}

\item{pattern}{A character string containing a expression to be matched
with all chromosomes files (\emph{e.g.}, "Chrom" for files which start by "Chrom"
followed by the chromosome number).}

\item{nSample}{The number of resampling done by \link{reSample} for p-values computation (minimum is 100).}

\item{empiricPvalue}{\code{empiricPvalue = TRUE} (default) compute PValue based on the null
distribution (resampling). If \code{empiricPvalue = TRUE}, the empirical p-values are computed instead.}

\item{mc.cores}{The number of cores to use (default is \code{1}),
\emph{i.e.}, at most how many child processes will be run simultaneously.
Must be at least one, and parallelization requires at least two cores.}

\item{onlyGenome}{\code{onlyGenome = TRUE} (default) compute resampling step for all chromosomes.}
}
\value{
Return a \code{list} of three elements:
\itemize{
\item object.xy \link{Enrichment} object from the comparison between \code{object.x} and \code{object.y}.
\item object.x \link{Enrichment} object passed in \code{object.x} with resampling data.
\item object.y \link{Enrichment} object passed in \code{object.y} with resampling data.
}
}
\description{
Compare the enrichment analysis between two set of SNPs.
\link{compareEnrichment} compare two \link{Enrichment} objects.
}
\note{
Still in development.
}
\examples{
if (interactive()) {
  data(toyEnrichment)

  reSample(
    object = toyEnrichment,
    nSample = 10,
    empiricPvalue = TRUE,
    MAFpool = c(0.05, 0.10, 0.2, 0.3, 0.4, 0.5),
    mc.cores = 1,
    onlyGenome = TRUE
  )

  excludeFile <- c(
    "rs7897180", "rs4725479", "rs315404", "rs17390391", "rs1650670",
    "rs6783390", "rs1642009", "rs4756586", "rs11995037", "rs4977345",
    "rs13136448", "rs4233536", "rs11151079", "rs2299657", "rs4833930",
    "rs1384", "rs7168184", "rs6909895", "rs7972667", "rs2293229",
    "rs918216", "rs6040608", "rs2817715", "rs13233541", "rs4486743",
    "rs2127806", "rs10912854", "rs1869052", "rs9853549", "rs448658",
    "rs2451583", "rs17483288", "rs10962314", "rs9612059", "rs1384182",
    "rs8049208", "rs12215176", "rs2980996", "rs1736976", "rs8089268",
    "rs10832329", "rs12446540", "rs7676237", "rs869922", "rs16823426",
    "rs1374393", "rs13268781", "rs11134505", "rs7325241", "rs7520109"
  )
  # OR
  excludeFile <- system.file("extdata/Exclude/toyExclude.txt", package = "snpEnrichment")

  toyEnrichment_exclude <- excludeSNP(toyEnrichment, excludeFile, mc.cores = 1)

  compareResults <- compareEnrichment(
    object.x = toyEnrichment,
    object.y = toyEnrichment_exclude,
    pattern = "Chrom",
    nSample = 10,
    empiricPvalue = FALSE,
    mc.cores = 1,
    onlyGenome = TRUE
  )
}

}
