% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{.checkTranscript}
\alias{.checkTranscript}
\title{.checkTranscript}
\usage{
.checkTranscript(data, transcriptFile, distThresh)
}
\arguments{
\item{distThresh}{}
}
\description{
.checkTranscript
}
\keyword{internal}
