% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{.compareEnrich}
\alias{.compareEnrich}
\title{.compareEnrich}
\usage{
.compareEnrich(object1, object2, nSample, empiricPvalue, sigThresh,
  MAFpool, mc.cores)
}
\arguments{
\item{mc.cores}{}
}
\description{
.compareEnrich
}
\keyword{internal}
