% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Enrichment-class}
\alias{Enrichment-class}
\alias{enrichment}
\alias{enrichment,ANY-method}
\title{An S4 class \linkS4class{Enrichment}}
\description{
This class is defined to summarize the enrichment analysis on each chromosomes and the whole genome.
}
\section{Slots}{

\describe{
\item{\code{Signal}}{A three columns \code{data.frame}: "SNP", "PVALUE" and "IN" (\emph{e.g.}, GWAS).
"IN" is computed during the reading step and gives informations about which SNPs are kept
for the enrichment analysis.}

\item{\code{Loss}}{A four columns data.frame: "Rows", "Unique", "Intersect.Ref.Signal" and "CIS".
This slot gives information on data losses.}

\item{\code{Call}}{Each parameters used for the reading or resampling step are stored in this slot.}

\item{\code{eSNP}}{Contain a \linkS4class{EnrichSNP} object for a list of SNPs (eSNP).}

\item{\code{xSNP}}{Contain a \linkS4class{EnrichSNP} object for a extended list of SNPs (xSNP).}

\item{\code{Chromosomes}}{A list of 22 \linkS4class{Chromosome} objects.}
}}

\section{Objects from the Class}{
 \linkS4class{Enrichment} is defined to build an object
of class \linkS4class{Enrichment} in order to compute an enrichment analysis.
\linkS4class{Enrichment} is the object containing the results for all \linkS4class{Chromosome} object
and for the whole genome.

When an \linkS4class{Enrichment} object is created, it contains a list of SNPs (\emph{e.g.}, eSNPs).
All the others slots are "empty".
After \link{reSample} is ran on an \linkS4class{Enrichment} object, the slots:
Table, EnrichmentRatio, Z, PValue and Resampling are filled.

Note that if \link{reSample} is executed on an \linkS4class{Enrichment} every new resampling is added to the
original ones, pre-existing statistics are erased and computed again with the new resampling set.
}

\examples{

data(toyEnrichment)
toyEnrich <- enrichment()
show(toyEnrich)

toyEnrich["Loss"] <- toyEnrichment["Loss"]
toyEnrich["Loss"]

toyEnrich <- enrichment(Loss = toyEnrichment["Loss"], eSNP = toyEnrichment["eSNP"])
toyEnrich <- enrichment(Loss = toyEnrichment["Loss"])

if (interactive()) {
  reSample(
    object = toyEnrichment,
    nSample = 10,
    empiricPvalue = TRUE,
    MAFpool = c(0.05, 0.10, 0.2, 0.3, 0.4, 0.5),
    mc.cores = 1,
    onlyGenome = TRUE
  )
  print(toyEnrichment)

  excludeFile <- c(
    "rs7897180", "rs4725479", "rs315404", "rs17390391", "rs1650670",
    "rs6783390", "rs1642009", "rs4756586", "rs11995037", "rs4977345",
    "rs13136448", "rs4233536", "rs11151079", "rs2299657", "rs4833930",
    "rs1384", "rs7168184", "rs6909895", "rs7972667", "rs2293229",
    "rs918216", "rs6040608", "rs2817715", "rs13233541", "rs4486743",
    "rs2127806", "rs10912854", "rs1869052", "rs9853549", "rs448658",
    "rs2451583", "rs17483288", "rs10962314", "rs9612059", "rs1384182",
    "rs8049208", "rs12215176", "rs2980996", "rs1736976", "rs8089268",
    "rs10832329", "rs12446540", "rs7676237", "rs869922", "rs16823426",
    "rs1374393", "rs13268781", "rs11134505", "rs7325241", "rs7520109"
  )

  toyEnrichment_exclude <- excludeSNP(toyEnrichment, excludeFile, mc.cores = 1)
  print(toyEnrichment_exclude)
}

}
