% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{EnrichSNP-class}
\alias{EnrichSNP-class}
\alias{EnrichSNP}
\alias{enrichSNP}
\alias{enrichSNP,ANY-method}
\title{Class \linkS4class{EnrichSNP}}
\description{
Class \linkS4class{EnrichSNP}

Constructor generic for \linkS4class{EnrichSNP}

Constructor method for \linkS4class{EnrichSNP}
}
\section{Slots}{

\describe{
\item{\code{List}}{A list of SNPs used to compute enrichment (\emph{e.g.}, eSNP or xSNP).}

\item{\code{Table}}{A contingency table with SNPs (columns) and P-Values from signal (rows).}

\item{\code{EnrichmentRatio}}{Enrichment Ratio is computed on the contingency table (\code{Table} slot).}

\item{\code{Z}}{A statistic computed from \code{EnrichmentRatio} and resampling results.}

\item{\code{PValue}}{P-Value associated with the statistic \code{Z}.}

\item{\code{Resampling}}{A matrix with by row, the contingency table and the odds ratio for each resampling.}
}}

\keyword{internal}
