% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{GC}
\alias{GC}
\title{GC}
\usage{
GC(verbose = getOption("verbose"), reset = FALSE)
}
\arguments{
\item{verbose}{A \code{logical}. If \code{TRUE}, the garbage collection prints statistics about cons cells
and the space allocated for vectors.}

\item{reset}{A \code{logical}. If \code{TRUE} the values for maximum space used are reset
to the current values.}
}
\description{
\code{GC} performs garbage collection until free memory indicators show no change.
}
\keyword{internal}
