% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFiles.R
\name{initFiles}
\alias{initFiles}
\title{Initialize files for enrichment analysis}
\usage{
initFiles(pattern = "Chrom", snpInfoDir, signalFile, mc.cores = 1)
}
\arguments{
\item{pattern}{A character string containing a expression to be matched with all
chromosomes files (\emph{e.g.},"Chrom" for files which start by "Chrom" followed
by the chromosome number).}

\item{snpInfoDir}{A character string naming a directory containing the reference data
in a PLINK format (.bed, .bim and .fam).}

\item{signalFile}{The name of the signal file which the data are to be read from
(2 columns: "SNP" and "PVALUE"). Each row of the table appears as one line of the file.
If it does not contain an \code{_absolute_} path, the file name is \code{_relative_} to the current
working directory, \link{getwd}. The fields separator character have to be a space \code{" "}
or a tabulation \code{"\t"}.}

\item{mc.cores}{The number of cores to use (default is \code{1}), \emph{i.e.},
at most how many child processes will be run simultaneously.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
This function writes several files, in the temporary directory (defined in \code{R_SESSION_TMPDIR}),
nothing else is returned. These files are used to build an \link{Enrichment} object by \link{readEnrichment}
in order to compute enrichment analysis (\link{reSample}).
}
\description{
\link{initFiles} create several files needed to run \link{readEnrichment}.
".frq" and ".signal" are created with PLINK.
LD computation can be run with \link{writeLD} or with PLINK.
}
\examples{
if (interactive()) {
  snpInfoDir <- system.file("extdata/snpInfo", package = "snpEnrichment")
  signalFile <- system.file("extdata/Signal/toySignal.txt", package = "snpEnrichment")
  initFiles(pattern = "Chrom", snpInfoDir, signalFile, mc.cores = 1)
}

}
