% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot}
\alias{plot}
\title{Plot method (S4) for \link{Enrichment} object}
\arguments{
\item{x}{An object of class \link{Enrichment} which the Z statistics or p-values have to be drawn.}

\item{what}{Default \code{"Genome"}. Plot Z statistics or p-values for genome only
(what must be: \code{"All"}, \code{"Genome"} or numeric vector).}

\item{type}{Plot the selected analysis for \code{"eSNP"} and/or \code{"xSNP"}.}

\item{ggplot}{Use ggplot (default \code{FALSE}) instead of classic plot method.}

\item{pvalue}{If \code{TRUE}, p-value convergense is plotted. Otherwise, Z statistic is plotted.}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see \link{par}).}
}
\description{
\link{plot} is a generic function for plotting of R objects.
The function invokes particular \link{methods} which depend on the \link{class} of the first argument.
}
\examples{
if (interactive()) {
  data(toyEnrichment)
  reSample(toyEnrichment, 10)
  plot(toyEnrichment)
}

}
