% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEnrichment.R
\name{readEnrichment}
\alias{readEnrichment}
\title{Read and create EnrichmentRatio object}
\usage{
readEnrichment(pattern = "Chrom", signalFile, transcriptFile = FALSE,
  snpListDir, snpInfoDir, distThresh = 1000, sigThresh = 0.05,
  LD = FALSE, ldDir = NULL, mc.cores = 1)
}
\arguments{
\item{pattern}{A character string containing a expression to be matched with all chromosomes files
(e.g."Chrom" for files which start by "Chrom" followed by the chromosome number).}

\item{signalFile}{The name of the signal file which the data are to be read from
(2 columns: "SNP" and "PVALUE"). Each row of the table appears as one line of the file.
If it does not contain an \code{_absolute_} path, the file name is \code{_relative_} to the current
working directory, \link{getwd}. The fields separator character have to be a space \code{" "}
or a tabulation \code{"\t"}.}

\item{transcriptFile}{A character string naming a file or a \link{data.frame} with four columns:
Chromomosome, trancript's name, Starting and Ending positions.
\code{data(trancript)} can be use as parameters. Default is \code{FALSE}.}

\item{snpListDir}{\link{character}: character string naming a directory
containing a list of SNPs for one or several chromosomes. \code{snpListDir}
can be a single file with at least two columns: chromosome and rs name.}

\item{snpInfoDir}{\link{character}: character string naming a directory
containing the reference data in a PLINK format (.bed, .bim and .fam).}

\item{distThresh}{\link{numeric}: maximal distance (kb) between SNP and gene.
\code{distThresh} is used if \code{transcriptFile} is set.}

\item{sigThresh}{\link{numeric}: statistical threshold for signal (\emph{e.g.}, \code{0.05} for a given GWAS signal)
used to compute an Enrichment Ratio.}

\item{LD}{\link{logical}: \code{LD=TRUE} (default is \code{FALSE}) read LD compute with \link{writeLD} function
or with PLINK. Note that, this setting can increase the computation's time,
depending on number of SNPs in the signal file.}

\item{ldDir}{\link{character}: character string naming a directory where the linkage disequilibrium files
should be read (default \code{NULL} is in temporary directory). LD files can be the LD output from plink.}

\item{mc.cores}{\link{numeric}: the number of cores to use (default is \code{1}),
\emph{i.e.}, at most how many child processes will be run simultaneously.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
Return an object of class \link{Enrichment} partly filled.
}
\description{
Read files created by \link{initFiles} and create an \link{Enrichment} object.
}
\examples{
if (interactive()) {
  snpListDir <- system.file("extdata/List", package = "snpEnrichment")
  signalFile <- system.file("extdata/Signal/toySignal.txt", package = "snpEnrichment")
  snpInfoDir <- system.file("extdata/snpInfo", package = "snpEnrichment")
  data(transcript)
  transcriptFile <- transcript

  initFiles(pattern = "Chrom", snpInfoDir, signalFile, mc.cores = 1)
  toyData <- readEnrichment(
    pattern = "Chrom",
    signalFile,
    transcriptFile,
    snpListDir,
    snpInfoDir,
    distThresh = 1000,
    sigThresh = 0.05,
    LD = FALSE,
    ldDir = NULL,
    mc.cores = 1
  )
  toyData
}

}
