% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\alias{reset,ANY-method}
\alias{reset,EnrichSNP-method}
\alias{reset,Chromosome-method}
\alias{reset,Enrichment-method}
\title{Reset methods.}
\arguments{
\item{oject}{An \linkS4class{Enrichment} object.}

\item{i}{A chracter naming the field to be reset.}
}
\description{
Reset any field of a \linkS4class{Enrichment} object to default (empty) values.
}
