\name{snpEnrichment-internal}
\alias{.verbose}
\alias{.showArgs}
\alias{.showArgs}
\alias{.writeSignal}
\alias{.writeFreq}
\alias{.readSignal}
\alias{.readSNP}
\alias{.verbose}
\alias{.checkFilePath}
\alias{.showArgs}
\alias{.writeSignal}
\alias{.writeFreq}
\alias{.readTranscript}
\alias{.readFreq}
\alias{.readLD}
\alias{.readFiles}
\alias{.enrichmentRatio}
\alias{.reSample}
\alias{.compareEnrich}
\alias{.SNP.show}
\alias{.Chromosome.show}
\alias{.Enrichment.show}

\alias{newSNP}
\alias{newSNP-methods}
\alias{newSNP,ANY-method}

\alias{is.SNP}
\alias{is.SNP-methods}
\alias{is.SNP,ANY-method}

\alias{reset}
\alias{reset-methods}
\alias{reset,Chromosome-method}
\alias{reset,Chromosome,ANY-method}
\alias{reset,Enrichment-method}
\alias{reset,Enrichment,ANY-method}
\alias{reset,ANY-method}

\alias{computeER}
\alias{computeER,Chromosome-method}
\alias{computeER,Enrichment-method}
\alias{computeER,ANY-method}
\alias{doLDblock}
\alias{doLDblock,Chromosome-method}
% \alias{doLDblock,ANY-method}


\title{~ Internal: snpEnrichment objects and methods ~}
\description{Internal: snpEnrichment objects and methods.}
\details{
These are not intended to be called by the user.\cr
If you have a specific need or for more details on snpEnrichment, feel free to contact the author.
}

\seealso{
Overview : \code{\link{snpEnrichment-package}} \cr
Classes : \code{\linkS4class{Enrichment}}, \code{\linkS4class{Chromosome}}, \code{\linkS4class{SNP}} \cr
Methods :  \code{\link{reset}}, \code{\link{reSample}}, \code{\link{excludeSNP}}, \code{\link{compareEnrichment}}, \cr
            \code{\link{enrichment}}, \code{\link{is.enrichment}}, \code{\link{chromosome}}, \code{\link{is.chromosome}} \cr
Functions : \code{\link{initFiles}}, \code{\link{writeLD}}, \code{\link{readEnrichment}}
}

\keyword{internal}