% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{toyEnrichment}
\alias{toyEnrichment}
\title{Toy dataset with SNP data}
\format{See class \linkS4class{Enrichment} for details about the format.}
\usage{
toyEnrichment
}
\description{
This data set gives an \linkS4class{Enrichment} object after, \link{initFiles} and \link{readEnrichment} is ran.
Compute LD for all SNPs in \code{snpListDir} files two by two. Genome Build 37.3 (hg19).
}
\keyword{datasets}
