\name{toyM1-dataset}
\docType{data}
\alias{toyM1}
\alias{toyM1-dataset}

\title{Toy dataset with SNP data}

\description{
This data set gives an \code{\linkS4class{Enrichment}} object after,
\code{\link{initFiles}} and \code{\link{readEnrichment}} is ran.
The extend method used is "LD" asssociation (in \code{\link{readEnrichment}}: \code{extendMethod="ld"}).
Genome Build 37.3 (hg19).
}

\usage{
data(toyM1)
toyM1
}

\format{
See class \code{\linkS4class{Enrichment}} for details about the format.
}

% \source{}

% \references{}

\keyword{datasets}