\name{toyM2-dataset}
\docType{data}
\alias{toyM2}
\alias{toyM2-dataset}

\title{Toy dataset with SNP data}

\description{
This data set gives an \code{\linkS4class{Enrichment}} object after,
\code{\link{initFiles}} and \code{\link{readEnrichment}} is ran.
The extend method used is "LD Block" asssociation (in \code{\link{readEnrichment}}: \code{extendMethod="block"}).
Genome Build 37.3 (hg19).
}

\usage{
data(toyM2)
toyM2
}

\format{
See class \code{\linkS4class{Enrichment}} for details about the format.
}

% \source{}

% \references{}

\keyword{datasets}