\name{transcript-dataset}
\docType{data}
\alias{transcript}
\alias{transcript-dataset}

\title{Transcript information in order to check the CIS status for SNPs}

\description{
This dataset is used by \code{\link{readEnrichment}} and \code{\link{compareEnrichment}} 
in order to check the CIS status for each SNP of signal.
Genome Build 37.3 (hg19).
}

\usage{
data(transcript)
transcript
}

\format{
See class \code{\link{readEnrichment}} and \code{\link{compareEnrichment}} for details about how to use this dataset.
}

% \source{}

% \references{}

\keyword{datasets}