% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{transcript}
\alias{transcript}
\title{Transcript information in order to check the CIS status for SNPs}
\format{See class \link{readEnrichment} and \link{compareEnrichment} for details about how to use this dataset.}
\description{
This dataset is used by \link{readEnrichment} and \link{compareEnrichment} in order to check the CIS status
for each SNP of signal. Genome Build 37.3 (hg19).
}
