% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLD.R
\name{writeLD}
\alias{writeLD}
\title{Linkage Disequilibrium (LD) computation with PLINK}
\usage{
writeLD(pattern = "Chrom", snpInfoDir, signalFile, ldDir = NULL,
  ldThresh = 0.8, depth = 1000, mc.cores = 1)
}
\arguments{
\item{pattern}{A character string containing a expression to be matched with all chromosomes files
(\emph{e.g.},\code{"Chrom"} for files which start by \code{"Chrom"} followed by the chromosome number).}

\item{snpInfoDir}{\link{character}: character string naming a directory
containing the reference data in a PLINK format (.bed, .bim and .fam).}

\item{signalFile}{The name of the signal file which the data are to be read from
(2 columns: "SNP" and "PVALUE"). Each row of the table appears as one line of the file.
If it does not contain an \code{_absolute_} path, the file name is \code{_relative_} to the current
working directory, \link{getwd}. The fields separator character have to be a space \code{" "}
or a tabulation \code{"\t"}.}

\item{ldDir}{\link{character}: character string naming a directory where the linkage disequilibrium files
should be read (default \code{NULL} is in temporary directory). LD files can be the LD output from plink.}

\item{ldThresh}{Threshold value for LD calculation.}

\item{depth}{This parameter is mandatory and controls the maximum lag between SNPs considered.}

\item{mc.cores}{\link{numeric}: the number of cores to use (default is \code{1}),
\emph{i.e.}, at most how many child processes will be run simultaneously.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
One ".ld" file per chromosome is returned by \link{writeLD} in \code{snpInfoDir} directory.
}
\description{
\link{writeLD} write a '.ld' file for each chromosomes which contains the LD (r^2).
}
\note{
The LD computation can take a long time depending on number of SNPs in
\code{signalFile}. It is recommended to save LD results in a directory
(\code{ldDir}) which is not a temporary directory.
}
\examples{
if (interactive()) {
  signalFile <- system.file("extdata/Signal/toySignal.txt", package = "snpEnrichment")
  snpInfoDir <- system.file("extdata/snpInfo", package = "snpEnrichment")
  writeLD(pattern = "Chrom", snpInfoDir, signalFile, ldDir = NULL, ldThresh = 0.8, mc.cores = 1)
}

}
