% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R, R/class.R
\name{internal}
\alias{internal}
\alias{.EnrichSNP.show}
\alias{.Chromosome.show}
\alias{.Enrichment.show}
\alias{.checkFilePath}
\alias{.checkSignalFile}
\alias{.checkSnpInfoDir}
\alias{.checkSnpListDir}
\alias{.checkTranscript}
\alias{.compareEnrich}
\alias{.enrichmentRatio}
\alias{.readFiles}
\alias{.readFreq}
\alias{.readLD}
\alias{.readSignal}
\alias{.readSNP}
\alias{.readTranscript}
\alias{.reSample}
\alias{.splitByChrom}
\alias{.verbose}
\alias{.writeFreq}
\alias{.writeSignal}
\alias{GC}
\alias{is.EnrichSNP}
\alias{is.chromosome}
\alias{is.enrichment}
\alias{[}
\alias{[,EnrichSNP-method}
\alias{[<-}
\alias{[<-,EnrichSNP-method}
\alias{print}
\alias{print,EnrichSNP-method}
\alias{show}
\alias{show,EnrichSNP-method}
\alias{[,Chromosome-method}
\alias{[<-,Chromosome-method}
\alias{print,Chromosome-method}
\alias{show,Chromosome-method}
\alias{[,Enrichment-method}
\alias{[<-,Enrichment-method}
\alias{print,Enrichment-method}
\alias{show,Enrichment-method}
\alias{computeER}
\alias{computeER-Chromosome}
\alias{computeER,Chromosome-method}
\alias{computeER-Enrichment}
\alias{computeER,Enrichment-method}
\alias{doLDblock}
\alias{doLDblock-Chromosome}
\alias{doLDblock,Chromosome-method}
\title{~ Internal: snpEnrichment objects and methods ~}
\usage{
.EnrichSNP.show(object)

.Chromosome.show(object)

.Enrichment.show(object)

.checkFilePath(path)

.checkSignalFile(signalFile)

.checkSnpInfoDir(snpInfoDir)

.checkSnpListDir(snpListDir, pattern)

.checkTranscript(data, transcriptFile, distThresh)

.compareEnrich(object1, object2, nSample, empiricPvalue, sigThresh,
  MAFpool, mc.cores)

.enrichmentRatio(table)

.readFiles(pattern, snpInfoDir, snpListDir, distThresh)

.readFreq(pattern, snpInfoDir)

.readLD(pattern, snpInfoDir, ldDir)

.readSignal(pattern)

.readSNP(pattern, snpListDir)

.readTranscript(transcriptFile)

.reSample(object, nSample, empiricPvalue, sigThresh, MAFpool, mc.cores)

.splitByChrom(pattern, snpListFile, directory)

.verbose(expr)

.writeFreq(pattern, snpInfoDir)

.writeSignal(pattern, snpInfoDir, signalFile)

GC(verbose = getOption("verbose"), reset = FALSE)

is.EnrichSNP(object)

is.chromosome(object)

is.enrichment(object)
}
\description{
These are not intended to be called by the user.
If you have a specific need or for more details on snpEnrichment, feel free to contact the author.
}
\keyword{internal}
