% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R, R/class.R
\name{internal}
\alias{internal}
\alias{.EnrichSNP.show}
\alias{.Chromosome.show}
\alias{.Enrichment.show}
\alias{.checkFilePath}
\alias{.checkSignalFile}
\alias{.checkSnpInfoDir}
\alias{.checkSnpListDir}
\alias{.checkTranscript}
\alias{.compareEnrich}
\alias{.enrichmentRatio}
\alias{.readFiles}
\alias{.readFreq}
\alias{.readLD}
\alias{.readSignal}
\alias{.readSNP}
\alias{.readTranscript}
\alias{.reSample}
\alias{.splitByChrom}
\alias{.verbose}
\alias{.writeFreq}
\alias{.writeSignal}
\alias{mclapply2}
\alias{maxCores}
\alias{GC}
\alias{is.EnrichSNP}
\alias{is.chromosome}
\alias{class-EnrichSNP}
\alias{EnrichSNP}
\alias{class-Chromosome}
\alias{Chromosome}
\alias{computeER}
\alias{computeER-Chromosome}
\alias{computeER-Enrichment}
\alias{doLDblock}
\alias{doLDblock-Chromosome}
\alias{reset}
\alias{reset-ANY}
\alias{reset-EnrichSNP}
\alias{reset-Chromosome}
\alias{reset-Enrichment}
\title{~ Internal: snpEnrichment objects and methods ~}
\usage{
.EnrichSNP.show(object)

.Chromosome.show(object)

.Enrichment.show(object)

.checkFilePath(path)

.checkSignalFile(signalFile)

.checkSnpInfoDir(snpInfoDir)

.checkSnpListDir(snpListDir, pattern)

.checkTranscript(data, transcriptFile, distThresh)

.compareEnrich(object1, object2, nSample, empiricPvalue, sigThresh,
  MAFpool, mc.cores)

.enrichmentRatio(table)

.readFiles(pattern, snpInfoDir, snpListDir, distThresh)

.readFreq(pattern, snpInfoDir)

.readLD(pattern, snpInfoDir, ldDir)

.readSignal(pattern)

.readSNP(pattern, snpListDir)

.readTranscript(transcriptFile)

.reSample(object, nSample, empiricPvalue, sigThresh, MAFpool, mc.cores)

.splitByChrom(pattern, snpListFile, directory)

.verbose(expr)

.writeFreq(pattern, snpInfoDir)

.writeSignal(pattern, snpInfoDir, signalFile)

mclapply2(X, FUN, ..., mc.preschedule = TRUE, mc.set.seed = TRUE,
  mc.silent = FALSE, mc.cores = getOption("mc.cores", 2L),
  mc.cleanup = TRUE, mc.allow.recursive = TRUE)

maxCores(mc.cores = 1)

GC(verbose = getOption("verbose"), reset = FALSE)

is.EnrichSNP(object)

is.chromosome(object)
}
\description{
These are not intended to be called by the user.
If you have a specific need or for more details on snpEnrichment, feel free to contact the author.
}
\section{Slots}{

\describe{
\item{\code{List}}{A list of SNPs used to compute enrichment (\emph{e.g.}, eSNP or xSNP).}

\item{\code{Table}}{A contingency table with SNPs (columns) and P-Values from signal (rows).}

\item{\code{EnrichmentRatio}}{Enrichment Ratio is computed on the contingency table (\code{Table} slot).}

\item{\code{Z}}{A statistic computed from \code{EnrichmentRatio} and resampling results.}

\item{\code{PValue}}{P-Value associated with the statistic \code{Z}.}

\item{\code{Resampling}}{A matrix with by row, the contingency table and the odds ratio for each resampling.}

\item{\code{Data}}{A \link{data.frame} with 6 columns (\code{"SNP"}, \code{"PVALUE"}, \code{"CHR"}, \code{"MAF"}, \code{"eSNP"}, \code{"xSNP"}).
Where \code{"eSNP"} and \code{"xSNP"} are logical columns defining the lists of SNPs (extended or not).}

\item{\code{LD}}{A data.frame which contains LD informations between SNPs (computed with \link{writeLD} or PLINK).}

\item{\code{eSNP}}{Contain a \link{EnrichSNP} object for a list of SNPs (eSNP).}

\item{\code{xSNP}}{Contain a \link{EnrichSNP} object for a extended list of SNPs (xSNP).}
}}

\keyword{internal}
