% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{reSample}
\alias{reSample}
\alias{reSample-Chromosome}
\alias{reSample-ANY}
\alias{reSample-Enrichment}
\title{Compute enrichment analysis on an \link{Enrichment} object}
\arguments{
\item{object}{An object to be updated. It is intended, an object returned by the \link{readEnrichment} function.}

\item{nSample}{The number of resampling done by \link{reSample} for p-values computation (minimum is 100).}

\item{empiricPvalue}{\code{TRUE} (default) compute PValue based on the null distribution (resampling).
If \code{TRUE}, the empirical p-values are computed instead.}

\item{sigThresh}{Statistical threshold for signal (\emph{e.g.},\code{0.05} for a given GWAS signal)
used to compute an Enrichment Ratio.}

\item{MAFpool}{Either a numeric vector giving the breaks points of intervals into which SNP's MAF
(Minor Allele Frequency) is to be split.}

\item{mc.cores}{The number of cores to use (default is \code{1}, \emph{i.e.}, at most how many child processes
will be run simultaneously. Must be at least one, and parallelization requires at least two cores.}

\item{onlyGenome}{Compute resampling step for all chromosomes (default \code{TRUE}).}
}
\value{
Return the object given in argument, updated by the resampling results.
}
\description{
After \link{initFiles} and \link{readEnrichment} has been run.
\link{reSample} computes a statistic value and a p-value for each chromosomes and for the whole genome.
}
\examples{
if (interactive()) {
  data(toyEnrichment)
  reSample(
    object = toyEnrichment,
    nSample = 10,
    empiricPvalue = TRUE,
    MAFpool = c(0.05, 0.10, 0.2, 0.3, 0.4, 0.5),
    onlyGenome = TRUE
  )
  toyEnrichment
}

}
