% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpEnrichment-package.R
\docType{package}
\name{snpEnrichment}
\alias{snpEnrichment}
\alias{snpEnrichment-package}
\title{~ Overview: SNPs Enrichment Analysis With Sampling ~}
\description{
Implements classes and methods for large scale SNP enrichment analysis 
    using a sampling approach (e.g. SNPs associated with genes expression in a GWAS signal).
}
\details{
Implements classes and methods for large-scale SNP enrichment analysis
(\emph{e.g.}, SNPs associated with genes expression in a GWAS signal).
}
\note{
Internal data management in \link{snpEnrichment} use RefSNP (rs) IDs.
}
\seealso{
\itemize{
\item Overview: \link{snpEnrichment}
\item Classes: \linkS4class{Enrichment} \linkS4class{Chromosome}
\item Methods:
\itemize{
\item \link{enrichment}
\item \link{chromosome}
\item \link{getEnrichSNP}
\item \link{plot}
\item \link{reSample}
\item \link{reset}
}
\item Functions:
\itemize{
\item \link{initFiles}
\item \link{writeLD}
\item \link{excludeSNP}
\item \link{readEnrichment}
\item \link{compareEnrichment}
}
}
}
\author{
\strong{Maintainer}: Mickaël Canouil \email{mickael.canouil@cnrs.fr}

Other contributors:
\itemize{
  \item Loic Yengo \email{loic.yengo@cnrs.fr} [contributor]
}

}
